+++ src/lib.rs
@@ -42,6 +42,9 @@ On server connected clients represented as entities with [`ConnectedClient`] com
 Their data represented as components, such as [`NetworkStats`]. Users can also attach their
 own metadata to them or even replicate these entiteis back to clients.
 
+These entities are spawned and despawned by the messaging backend. To request
+a disconnection, use the [`DisconnectRequest`] event.
+
 You can use [`Trigger<OnAdd, ConnectedClient>`] to react to new connections,
 or use backend-provided events if you need the disconnect reason.
 
@@ -619,6 +622,7 @@ pub mod prelude {
         shared::{
             RepliconSharedPlugin, SERVER,
             backend::{
+                DisconnectRequest,
                 connected_client::{ConnectedClient, NetworkStats},
                 replicon_channels::{Channel, RepliconChannels},
                 replicon_client::{RepliconClient, RepliconClientStatus},